[UPDATE] kds_feature AdoptguideActivity.kt and GuriguideActivity.kt
