+++ app/src/main/java/com/example/mungandnyang/AdoptguideActivity.kt
@@ -22,82 +22,23 @@ import com.google.android.gms.maps.model.CameraPosition
 import com.google.android.gms.maps.model.LatLng
 import com.google.android.gms.maps.model.MarkerOptions
 
-class AdoptguideActivity : AppCompatActivity() , GoogleApiClient.ConnectionCallbacks,
-    GoogleApiClient.OnConnectionFailedListener, OnMapReadyCallback {
-    companion object{
-        const val LATITUDE = 37.501288
-        const val LONGTITUDE = 126.878581
-        const val ADOPTNAME = "서울동물복지지원 구리센터"
-    }
-    lateinit var binding: ActivityGuriguideBinding
-    var apiClient: GoogleApiClient? = null
+class AdoptguideActivity : AppCompatActivity()  {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        binding = ActivityGuriguideBinding.inflate(layoutInflater)
+        val binding = ActivityAdoptguideBinding.inflate(layoutInflater)
         setContentView(binding.root)
-        val requestPermissionLauncher = registerForActivityResult(
-            ActivityResultContracts.RequestMultiplePermissions()){
-            if(it.all { permission -> permission.value == true }){
-                apiClient?.connect()
-            }else{
-                Toast.makeText(this, "권한거부 이유로 맵앱을 사용할 수 없습니다", Toast.LENGTH_SHORT).show()
-                finish()
-            }
-        }
 
-        (supportFragmentManager.findFragmentById(R.id.guriMapView) as SupportMapFragment?)?.getMapAsync(this)
-
-        if(
-            ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_FINE_LOCATION) !== PackageManager.PERMISSION_GRANTED ||
-            ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_COARSE_LOCATION) !== PackageManager.PERMISSION_GRANTED ||
-            ContextCompat.checkSelfPermission(this, Manifest.permission.ACCESS_NETWORK_STATE) !== PackageManager.PERMISSION_GRANTED ||
-            ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) !== PackageManager.PERMISSION_GRANTED ){
-
-            requestPermissionLauncher.launch( arrayOf(
-                Manifest.permission.ACCESS_FINE_LOCATION,
-                Manifest.permission.ACCESS_COARSE_LOCATION,
-                Manifest.permission.ACCESS_NETWORK_STATE,
-                Manifest.permission.WRITE_EXTERNAL_STORAGE
-            ))
-        }else{
-            apiClient?.connect()
-        }
-
-        binding.ivCall.setOnClickListener{
-            val intent = Intent(Intent.ACTION_DIAL)
-            intent.data = Uri.parse("tel:02-2636-7645")
+        binding.btnMapo.setOnClickListener{
+            val intent = Intent(this, MapoguideActivity::class.java)
             startActivity(intent)
+            finish()
         }
-    }
 
-    override fun onConnected(p0: Bundle?) {
-    }
-
-    override fun onConnectionSuspended(data: Int) {
-        Log.d("mungmap", "Location Provider 더 이상 이용이 불가능한 상황 ${data}")
-    }
-
-    override fun onConnectionFailed(connectionResult: ConnectionResult) {
-        Log.d("mungmap", "Location Provider가 제공되지 않는 상황 ${connectionResult.errorMessage}")
-    }
-
-    override fun onMapReady(googleMap: GoogleMap) {
-        val center = LatLng(LATITUDE, LONGTITUDE)
-        googleMap.addMarker(MarkerOptions().position(center).title(ADOPTNAME))
-        googleMap.moveCamera(CameraUpdateFactory.newLatLng(center))
-
-        val cameraPosition = CameraPosition.Builder()
-            .target(center)
-            .zoom(14f)
-            .build()
-        googleMap.moveCamera(CameraUpdateFactory.newCameraPosition(cameraPosition))
-    }
-
-    override fun onBackPressed() {
-        super.onBackPressed()
-        val intent = Intent(this, AdoptguideActivity::class.java)
-        startActivity(intent)
-        finish()
+        binding.btnGuri.setOnClickListener{
+            val intent = Intent(this, GuriguideActivity::class.java)
+            startActivity(intent)
+            finish()
+        }
     }
 }
